+++ apps/web/src/App.tsx
@@ -1,4 +1,4 @@
-import { FavolinkProvider, setAxiosUrl } from '@favolink-ui/react';
+import { FavolinkProvider } from '@favolink-ui/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { RouterProvider } from 'react-router-dom';
@@ -6,8 +6,6 @@ import { router } from './router';
 import '@favolink-ui/react/reset.css';
 import '@favolink-ui/react/theme.css';
 import '@/styles/font.css';
-
-setAxiosUrl(import.meta.env.VITE_API_HOST as string);
 
 const queryClient = new QueryClient({
   defaultOptions: {
