+++ packages/ui/react/package.json
@@ -13,7 +13,6 @@
   "dependencies": {
     "@favolink-ui/hooks": "workspace:*",
     "@favolink-ui/icons": "workspace:*",
-    "@favolink-ui/utils": "workspace:*",
     "@vanilla-extract/css": "^1.14.1",
     "@vanilla-extract/dynamic": "^2.1.0",
     "@vanilla-extract/recipes": "^0.5.3",
