+++ packages/ui/react/src/components/input/input-element.tsx
@@ -1,6 +1,6 @@
-import { cx } from '@favolink-ui/utils';
 import * as styles from './input-element.styles.css';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx } from '../../utils';
 
 type InputElementProps = HTMLFavolinkProps<'div'>;
