+++ packages/ui/react/src/components/menu/menu-trigger.tsx
@@ -1,6 +1,6 @@
-import { composeRefs, cx, mergeFns } from '@favolink-ui/utils';
 import { useMenuContext } from './menu.context';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { composeRefs, cx, mergeFns } from '../../utils';
 
 export type MenuTriggerProps = HTMLFavolinkProps<'button'>;
