+++ packages/ui/react/src/components/modal/modal-body.tsx
@@ -1,5 +1,5 @@
-import { cx } from '@favolink-ui/utils';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx } from '../../utils';
 
 export type ModalBodyProps = HTMLFavolinkProps<'main'>;
