+++ packages/ui/react/src/components/modal/modal-content.tsx
@@ -1,6 +1,6 @@
-import { cx } from '@favolink-ui/utils';
 import * as styles from './modal-content.styles.css';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx } from '../../utils';
 
 export type ModalContentProps = HTMLFavolinkProps<'div'>;
