+++ packages/ui/react/src/components/modal/modal-footer.tsx
@@ -1,6 +1,6 @@
-import { cx } from '@favolink-ui/utils';
 import * as styles from './modal-footer.styles.css';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx } from '../../utils';
 
 export type ModalFooterProps = HTMLFavolinkProps<'footer'>;
