+++ packages/ui/react/src/components/modal/modal-title.tsx
@@ -1,6 +1,6 @@
-import { cx } from '@favolink-ui/utils';
 import * as styles from './modal-title.styles.css';
 import { forwardRef } from '../../system';
+import { cx } from '../../utils';
 import { Heading, type HeadingProps } from '../typography';
 
 export type ModalTitleProps = HeadingProps & styles.ModalTitleVariants;
