+++ packages/ui/react/src/components/modal/modal-topbar.tsx
@@ -1,9 +1,9 @@
 import { ChevronLeftIcon, CloseIcon } from '@favolink-ui/icons';
-import { cx, mergeFns } from '@favolink-ui/utils';
 import { type MouseEvent } from 'react';
 import * as styles from './modal-topbar.styles.css';
 import { useModalContext } from './modal.context';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx, mergeFns } from '../../utils';
 
 export type ModalTopbarProps = HTMLFavolinkProps<'div'> &
   styles.ModalTopbarVariants & {
