+++ packages/ui/react/src/components/modal/modal-trigger.tsx
@@ -1,6 +1,6 @@
-import { cx, mergeFns } from '@favolink-ui/utils';
 import { useModalContext } from './modal.context';
 import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
+import { cx, mergeFns } from '../../utils';
 
 export type ModalTriggerProps = HTMLFavolinkProps<'button'>;
