--- packages/ui/utils/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@favolink-ui/utils",
-  "version": "0.0.0",
-  "private": true,
-  "main": "src/index.ts",
-  "scripts": {
-    "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
-  },
-  "dependencies": {
-    "axios": "^1.6.7",
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.3.3"
-  }
-}
